# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.1"
  title:  Hello World App
# during dev, should point to your local machine
host: my-americano-api.appspot.com
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /footballteams:
    x-swagger-router-controller: footballteams.controller
    get:
      description: get the football teams list
      operationId: getFootballteams
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetfootballteamsListResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse" 

    post:
      description: add a new Football teams to the list
      operationId: addFootballteams
      parameters:
        - in: body
          name: footballteams
          description: The Football Teams to create.
          schema:
            $ref: "#/definitions/Footballteams"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetfootballteamsResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /footballteams/{id}:
    x-swagger-router-controller: footballteams.controller
    get:
      description: get a football team
      operationId: getfootballteamsbyId
      parameters:
        - name: id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetfootballteamsResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update a football team
      operationId: updatefootballteams
      # define the parameters
      parameters:
        - name: id
          description: football team id
          type: number
          in: path
          required: true
        - in: body
          name: footballteams
          description: The football teams to update.
          schema:
            $ref: "#/definitions/Footballteams"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetfootballteamsResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
    delete:
      description: delete a football team
      operationId: deletefootballteams
      # define the parameters
      parameters:
        - name: id
          description: football team id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"    
   
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Footballteams:
    type: object
    properties:
      Name:
        type: string
        description: Name of the Football Teams
      City:
        type: string
        description: City from of the city
      Conference:
        type: string
        description: Conference to which the team belongs
      Division:
        type: string
        description: City to which the team belongs     
       
  GetfootballteamsResponse:
    type: object
    properties:
      id:
        type: number
        description: identifier
      name:
        type: string
        description: Name of the Football Teams
      city:
        type: string
        description: City from of the city
      Conference:
        type: string
        description: Conference to which the team belongs
      Division:
        type: string
        description: City to which the team belongs  
         
  GetfootballteamsListResponse:
    required:
      - footballteams
    properties:
      footballteams:
        type: array
        items:
          $ref: "#/definitions/GetfootballteamsResponse"
          
          
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
     
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        